packages/channels/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ninetrix-channels"
version = "0.1.0"
description = "Pluggable inbound channel adapters for Ninetrix agents"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
]

[project.optional-dependencies]
discord = ["discord.py>=2.3"]
dev = ["pytest", "pytest-asyncio"]

[tool.setuptools.packages.find]
include = ["ninetrix_channels*"]
